optimize hashing
This patch includes two fixes for a roughly 30% speed-up.

The first one is to use sip13's optimized string hashing function.  This
removes the need to do extra string->[]byte conversions (malloc+copy)
every time we want to hash something.

The second is to use multiplication as an alternative to modulo
reduction.